requirements.txt
click
openai
python-dotenv
agentfs-sdk